+++ Source/NSButton+PecuniaAdditions.m
@@ -25,7 +25,7 @@
 - (NSColor *)textColor
 {
     NSAttributedString *attrTitle = [self attributedTitle];
-    int                len = [attrTitle length];
+    NSUInteger         len = [attrTitle length];
     NSRange            range = NSMakeRange(0, MIN(len, 1)); // take color from first char
     NSDictionary       *attrs = [attrTitle fontAttributesInRange: range];
     NSColor            *textColor = [NSColor controlTextColor];
@@ -39,7 +39,7 @@
 {
     NSMutableAttributedString *attrTitle = [[NSMutableAttributedString alloc]
                                             initWithAttributedString: [self attributedTitle]];
-    int     len = [attrTitle length];
+    NSUInteger len = [attrTitle length];
     NSRange range = NSMakeRange(0, len);
     [attrTitle addAttribute: NSForegroundColorAttributeName
                       value: textColor
@@ -55,7 +55,7 @@
 - (NSColor *)textColor
 {
     NSAttributedString *attrTitle = [self attributedTitle];
-    int                len = [attrTitle length];
+    NSUInteger         len = [attrTitle length];
     NSRange            range = NSMakeRange(0, MIN(len, 1)); // take color from first char
     NSDictionary       *attrs = [attrTitle fontAttributesInRange: range];
     NSColor            *textColor = [NSColor controlTextColor];
@@ -69,7 +69,7 @@
 {
     NSMutableAttributedString *attrTitle = [[NSMutableAttributedString alloc]
                                             initWithAttributedString: [self attributedTitle]];
-    int     len = [attrTitle length];
+    NSUInteger len = [attrTitle length];
     NSRange range = NSMakeRange(0, len);
     [attrTitle addAttribute: NSForegroundColorAttributeName
                       value: textColor
